[refactoring] split header files for all command
