+++ Shell/Erase.cpp
@@ -1,27 +1,8 @@
 #pragma once
 
-#include "CommandHandler.h"
-#include "LbaRangeVerifier.h"
+#include "Erase.h"
 
-#include <iostream>
-
-using namespace std;
-class Erase : public CommandHandler
-{
-public:
-	Erase(ostream& _out, SsdHelper& _ssd) : CommandHandler(_out, _ssd) {};
-	bool isValidArgs(const vector<string>& args) override;
-
-	Progress doCommand(const vector<string>& args) override;
-
-	void usage() override {};
-
-	~Erase() {};
-private:
-	LbaRangeVerifier m_lbaRangeVerifier;
-};
-
-inline bool Erase::isValidArgs(const vector<string>& args)
+bool Erase::isValidArgs(const vector<string>& args)
 {
 	if (args.size() != 3)
 		return INVALID;
@@ -33,7 +14,7 @@ inline bool Erase::isValidArgs(const vector<string>& args)
 	return VALID;
 }
 
-inline Progress Erase::doCommand(const vector<string>& args)
+Progress Erase::doCommand(const vector<string>& args)
 {
 	logger.print("Command : " + sliceString(args, 0));
 	string arguments = "E " + args[1] + " " + args[2];
