+++ Shell/EraseRange.cpp
@@ -1,27 +1,8 @@
 #pragma once
 
-#include "CommandHandler.h"
-#include "LbaRangeVerifier.h"
+#include "EraseRange.h"
 
-#include <iostream>
-
-using namespace std;
-class EraseRange : public CommandHandler
-{
-public:
-	EraseRange(ostream& _out, SsdHelper& _ssd) : CommandHandler(_out, _ssd) {};
-	bool isValidArgs(const vector<string>& args) override;
-
-	Progress doCommand(const vector<string>& args) override;
-
-	void usage() override {};
-
-	~EraseRange() {};
-private:
-	LbaRangeVerifier m_lbaRangeVerifier;
-};
-
-inline bool EraseRange::isValidArgs(const vector<string>& args)
+bool EraseRange::isValidArgs(const vector<string>& args)
 {
 	if (args.size() != 3)
 		return INVALID;
@@ -33,7 +14,7 @@ inline bool EraseRange::isValidArgs(const vector<string>& args)
 	return VALID;
 }
 
-inline Progress EraseRange::doCommand(const vector<string>& args)
+Progress EraseRange::doCommand(const vector<string>& args)
 {
 	logger.print("Command : " + sliceString(args, 0));
 	auto numberOfLba = stoi(args[2]) - stoi(args[1]) + 1;
