+++ Shell/FullWrite.h
@@ -1,0 +1,26 @@
+#pragma once
+
+#include <string>
+#include <iostream>
+
+#include "Write.h"
+#include "CommandHandler.h"
+
+using namespace std;
+
+class FullWrite : public CommandHandler
+{
+public:
+	FullWrite(ostream& _out, SsdHelper& _ssd, Write* _write) : CommandHandler(_out, _ssd), write(_write){};
+
+	bool isValidArgs(const vector<string>& args) override;
+
+	Progress doCommand(const vector<string>& args) override;
+
+	void usage() override {}
+
+	~FullWrite() {};
+private:
+	Write* write;
+	vector<vector<string>> nArgs;
+};
